Refactor Menu to avoid looping methods
