+++ src/main/java/com/infoshareacademy/App.java
@@ -1,12 +1,12 @@
 package com.infoshareacademy;
 
-import com.infoshareacademy.menu.MenuNavigation;
+import com.infoshareacademy.menu.MenuControl;
 
 public class App {
 
     public static void main(String[] args) {
-        MenuNavigation menuNavigation = new MenuNavigation();
-        menuNavigation.mainNavigation();
+        MenuControl menuControl = new MenuControl();
+        menuControl.mainNavigation();
 
     }
 }
