+++ src/main/java/com/infoshareacademy/menu/UserInput.java
@@ -1,23 +1,22 @@
 package com.infoshareacademy.menu;
 
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+
 import java.util.Scanner;
 
 public class UserInput {
 
+    private static final Logger STDOUT = LoggerFactory.getLogger("CONSOLE_OUT");
+    private static final String CHOICE_MESSAGE = "\nType your choice: ";
+
     Scanner scanner = new Scanner(System.in);
 
     public int getUserInput() {
-        System.out.print("Type your choice: ");
+        STDOUT.info(CHOICE_MESSAGE);
         String userChoice = scanner.nextLine();
 
-        Integer userChoiceParsed = -1; //value not available in any menu
-        try {
-            userChoiceParsed = Integer.parseInt(userChoice);
-        } catch (NumberFormatException numberFormatException) {
-            STDOUT.info("Wrong input format! Only numbers accepted. Try again. ");
-            Utilities.freezeConsole(1);
-        }
-        return userChoiceParsed;
+        return Integer.parseInt(userChoice);
     }
 
 }
